+++ packages/tts-ggml/addonLogging.d.ts
@@ -2,7 +2,7 @@ export interface AddonLogging {
     setLogger(this: void, callback: (priority: number, message: string) => void): void;
     releaseLogger(this: void): void;
 }
-export declare const setLogger: (this: void, callback: (priority: number, message: string) => void) => void;
-export declare const releaseLogger: (this: void) => void;
+export declare const setLogger: AddonLogging["setLogger"];
+export declare const releaseLogger: AddonLogging["releaseLogger"];
 declare const addonLogging: AddonLogging;
 export default addonLogging;